This is a new version of save_acquire
This new version adds some automation to save_acquire. This will create a new folder within the 'Date' directory each time 'save_acquire' is run. Additionally, the saving of intensity image JPGs seems unnecessary, as we are only post processing the numpy arrays at this point so the JPG saving process was removed. This is saved on the lab computer as save_acquire_revised_191216_kjl.py.