New version of data_1 with more automation
This version no longer requires user to know the number of numpy arrays in the folder which their data is saved. Furthermore, the data is more modular, having each data run from save_acquire saved to a different folder. The phase map generated in this script also has a time stamp now in case multiple phase maps of the same data are ever made.