dark.py revised (4D Dark Noise and saving)
After discussing the over averaging issue with Paul, Cole and I went ahead with Mauricio's suggestion to do histograms of individual pixels for individual (exp,gain) settings. Due to lack of time, instead of resolving conflicts with my other current branch, I created a new branch (which if this code works when ran on lab computer with camera) would be merged with the master. Note, PixelPoints function is left defined but unused in this dark.py version. More user defined inputs also.