2D proof-of-concept MCMC optimization method
This is an unpolished version of what I have been working on recently to demonstrate an MCMC method for optimization of a toy 2D parameter space and analytical function. NOTE: currently the print statements are rather long and unwieldy when the NumPaths is large (say, 100) but useful for evaluating the effectiveness of the method for small NumPaths. The print statements will be removed in future uploads.